Use only CEL validation for Compositions
After removing (MR) schema aware validation, everything that's left
could be done with just CEL rules.
