+++ cluster/webhookconfigurations/manifests.yaml
@@ -24,23 +24,3 @@ webhooks:
     resources:
     - compositeresourcedefinitions
   sideEffects: None
-- admissionReviewVersions:
-  - v1
-  clientConfig:
-    service:
-      name: webhook-service
-      namespace: system
-      path: /validate-apiextensions-crossplane-io-v1-composition
-  failurePolicy: Fail
-  name: compositions.apiextensions.crossplane.io
-  rules:
-  - apiGroups:
-    - apiextensions.crossplane.io
-    apiVersions:
-    - v1
-    operations:
-    - UPDATE
-    - CREATE
-    resources:
-    - compositions
-  sideEffects: None
