+++ cmd/crank/render/cmd.go
@@ -151,14 +151,6 @@ func (c *Cmd) Run(k *kong.Context, log logging.Logger) error { //nolint:gocognit
 		return errors.Errorf("composition's compositeTypeRef.apiVersion (%s) does not match XR's apiVersion (%s)", compRef.APIVersion, xrGVK.GroupVersion().String())
 	}
 
-	warns, errs := comp.Validate()
-	for _, warn := range warns {
-		_, _ = fmt.Fprintf(k.Stderr, "WARN(composition): %s\n", warn)
-	}
-	if len(errs) > 0 {
-		return errors.Wrapf(errs.ToAggregate(), "invalid Composition %q", comp.GetName())
-	}
-
 	// check if XR's matchLabels have corresponding label at composition
 	xrSelector := xr.GetCompositionSelector()
 	if xrSelector != nil {
