--- pkg/validation/apiextensions/v1/composition/testdata/complex_schema_openidconnectproviders_v1beta1.json
@@ -1,1 +1,0 @@
-{"properties":{"apiVersion":{"type":"string"},"kind":{"type":"string"},"metadata":{"type":"object"},"spec":{"properties":{"deletionPolicy":{"default":"Delete","enum":["Orphan","Delete"],"type":"string"},"forProvider":{"properties":{"clientIDList":{"items":{"type":"string"},"maxItems":100,"type":"array"},"tags":{"items":{"properties":{"key":{"type":"string"},"value":{"type":"string"}},"required":["key"],"type":"object"},"type":"array"},"thumbprintList":{"items":{"type":"string"},"maxItems":5,"minItems":1,"type":"array"},"url":{"type":"string"}},"required":["thumbprintList","url"],"type":"object"},"providerConfigRef":{"default":{"name":"default"},"properties":{"name":{"type":"string"},"policy":{"properties":{"resolution":{"default":"Required","enum":["Required","Optional"],"type":"string"},"resolve":{"enum":["Always","IfNotPresent"],"type":"string"}},"type":"object"}},"required":["name"],"type":"object"},"providerRef":{"properties":{"name":{"type":"string"},"policy":{"properties":{"resolution":{"default":"Required","enum":["Required","Optional"],"type":"string"},"resolve":{"enum":["Always","IfNotPresent"],"type":"string"}},"type":"object"}},"required":["name"],"type":"object"},"publishConnectionDetailsTo":{"properties":{"configRef":{"default":{"name":"default"},"properties":{"name":{"type":"string"},"policy":{"properties":{"resolution":{"default":"Required","enum":["Required","Optional"],"type":"string"},"resolve":{"enum":["Always","IfNotPresent"],"type":"string"}},"type":"object"}},"required":["name"],"type":"object"},"metadata":{"properties":{"annotations":{"additionalProperties":{"type":"string"},"type":"object"},"labels":{"additionalProperties":{"type":"string"},"type":"object"},"type":{"type":"string"}},"type":"object"},"name":{"type":"string"}},"required":["name"],"type":"object"},"writeConnectionSecretToRef":{"properties":{"name":{"type":"string"},"namespace":{"type":"string"}},"required":["name","namespace"],"type":"object"}},"required":["forProvider"],"type":"object"},"status":{"properties":{"atProvider":{"properties":{"createDate":{"format":"date-time","type":"string"}},"type":"object"},"conditions":{"items":{"properties":{"lastTransitionTime":{"format":"date-time","type":"string"},"message":{"type":"string"},"reason":{"type":"string"},"status":{"type":"string"},"type":{"type":"string"}},"required":["lastTransitionTime","reason","status","type"],"type":"object"},"type":"array"}},"type":"object"}},"required":["spec"],"type":"object"}
