+++ test/e2e/manifests/apiextensions/composition/validation/composition-invalid-empty-pipeline.yaml
@@ -1,0 +1,10 @@
+apiVersion: apiextensions.crossplane.io/v1
+kind: Composition
+metadata:
+  name: valid
+spec:
+  compositeTypeRef:
+    apiVersion: nop.example.org/v1alpha1
+    kind: XNopResource
+  mode: Pipeline
+  pipeline: []
