+++ test/e2e/manifests/apiextensions/composition/validation/composition-invalid-missing-secretref.yaml
@@ -14,6 +14,3 @@ spec:
     credentials:
     - name: some-secret
       source: Secret
-      secretRef:
-        namespace: default
-        name: some-secret
