--- test/e2e/manifests/apiextensions/composition/validation/composition-transform-tojson-valid.yaml
@@ -1,33 +1,0 @@
-apiVersion: apiextensions.crossplane.io/v1
-kind: Composition
-metadata:
-  name: valid-tojson-patch
-  annotations:
-    crossplane.io/composition-schema-aware-validation-mode: strict
-spec:
-  compositeTypeRef:
-    apiVersion: nop.example.org/v1alpha1
-    kind: XNopResource
-  resources:
-  - name: nop-resource-1
-    base:
-     apiVersion: nop.crossplane.io/v1alpha1
-     kind: NopResource
-     spec:
-      forProvider:
-        conditionAfter:
-        - conditionType: Ready
-          conditionStatus: "False"
-          time: 0s
-        - conditionType: Ready
-          conditionStatus: "True"
-          time: 1s
-    patches:
-    - type: FromCompositeFieldPath
-      fromFieldPath: spec
-      toFieldPath: metadata.annotations[spec]
-      transforms:
-      - type: string
-        string:
-         type: Convert
-         convert: ToJson
