--- test/e2e/manifests/apiextensions/composition/validation/setup/definition.yaml
@@ -1,36 +1,0 @@
-apiVersion: apiextensions.crossplane.io/v1
-kind: CompositeResourceDefinition
-metadata:
-  name: xnopresources.nop.example.org
-spec:
-  group: nop.example.org
-  names:
-    kind: XNopResource
-    plural: xnopresources
-  claimNames:
-    kind: NopResource
-    plural: nopresources
-  connectionSecretKeys:
-  - test
-  versions:
-  - name: v1alpha1
-    served: true
-    referenceable: true
-    schema:
-     openAPIV3Schema:
-       type: object
-       properties:
-        spec:
-          type: object
-          properties:
-            coolField:
-              type: string
-          required:
-          - coolField
-        status:
-          type: object
-          properties:
-            coolerField:
-              type: string
-            evenCoolerField:
-              type: integer
