--- test/e2e/manifests/apiextensions/composition/validation/setup/provider.yaml
@@ -1,7 +1,0 @@
-apiVersion: pkg.crossplane.io/v1
-kind: Provider
-metadata:
-  name: provider-nop
-spec:
-  package: xpkg.upbound.io/crossplane-contrib/provider-nop:v0.3.0
-  ignoreCrossplaneConstraints: true
